Remove player save timestamp from DB schema
Checking against the save timestamp is not a solution to the concurrent
player saving problem because SystemTime is not monotonic, so saving old
player state would still be possible under a race condition. The proper
solution is to enqueue all player save operations with the DB mutex, but
this can block for a while under the new model since the op queue needs
to be flushed when a sync operation is invoked.
In a perfect world, all DB operations are async, but that's not easy.
It's a problem for another day.