"Gateway" server behavior
On shard server shutdown, send a shard warp packet to all game clients
with the shard ID. The login server will queue connection requests. If
the shard comes back up within the timeout period, the client will
automatically be reconnected.
This approach has the advantage of not requiring an actual separate
gateway server to manage connections. However, it only works on
graceful shutdown, but that should be the case most of the time.